Remove one layer of modularization and improve screen size responsiveness
