+++ frontend/src/components/config/App.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { Switch } from 'react-router-dom';
-import MainLayout from '../../layout/Main';
-import LandingPage from '../../../views/Landing';
-import NotFound from '../../../views/NotFound';
-import { CustomRoute } from '../Route';
-import GamePage from '../../../views/Game';
-import GameLayout from '../../layout/Game';
-import CreateGamePage from '../../../views/Create';
-import JoinGamePage from '../../../views/Join';
-import GameResultsPage from '../../../views/Results';
+import MainLayout from '../layout/Main';
+import LandingPage from '../../views/Landing';
+import NotFound from '../../views/NotFound';
+import { CustomRoute } from './Route';
+import GamePage from '../../views/Game';
+import GameLayout from '../layout/Game';
+import CreateGamePage from '../../views/Create';
+import JoinGamePage from '../../views/Join';
+import GameResultsPage from '../../views/Results';
 
 function App() {
   return (
