+++ frontend/src/components/core/Container.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { StyledMainContainer } from './styles';
+import styled from 'styled-components';
 
 type MyProps = {
   children: React.ReactNode,
@@ -9,5 +9,11 @@ const createContainer = (Container: any, props: MyProps) => {
   const { children } = props;
   return <Container>{children}</Container>;
 };
+
+const StyledMainContainer = styled.div`
+  margin: 0 auto;
+  padding: 10vw 0;
+  width: 80%;
+`;
 
 export const MainContainer = (props: MyProps) => createContainer(StyledMainContainer, props);
