--- frontend/src/components/core/Container/styles.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledMainContainer = styled.div`
-  margin: 0 auto;
-  padding: 10vw 0;
-  width: 80%;
-`;
