+++ frontend/src/components/core/Text.tsx
