+++ frontend/src/components/layout/Game.tsx
