+++ frontend/src/index.tsx
@@ -10,12 +10,18 @@ import Theme from './components/config/Theme';
 const GlobalStyle = createGlobalStyle`
   html {
     height: 100vh;
+    min-width: 640px;
     // @ts-ignore
     font-family: ${({ theme }) => theme.font};
     // @ts-ignore
-    font-size: ${({ theme }) => theme.fontSize.global};
+    font-size: ${({ theme }) => theme.fontSize.globalDefault};
     // @ts-ignore
     color: ${({ theme }) => theme.colors.text};
+    
+    @media (max-width: 800px) {
+      // @ts-ignore
+      font-size: ${({ theme }) => theme.fontSize.globalSmall};
+    }
   }
   
   body {
