+++ frontend/src/views/Results.tsx
